feat: :lipstick: Unified Terminal
