+++ gui/src/components/mainInput/belowMainInput/ThinkingBlockPeek.tsx
@@ -5,38 +5,10 @@ import { ChatHistoryItem } from "core";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 
-import { defaultBorderRadius, lightGray, vscBackground } from "../..";
-import { getFontSize } from "../../../util";
+import { vscBackground } from "../..";
 import { AnimatedEllipsis } from "../../AnimatedEllipsis";
 import StyledMarkdownPreview from "../../StyledMarkdownPreview";
-
-const SpoilerButton = styled.div`
-  background-color: ${vscBackground};
-  width: fit-content;
-  margin: 8px 6px 0px 2px;
-  font-size: ${getFontSize() - 2}px;
-  border: 0.5px solid ${lightGray};
-  border-radius: ${defaultBorderRadius};
-  padding: 4px 8px;
-  color: ${lightGray};
-  cursor: pointer;
-  box-shadow:
-    0 4px 6px rgba(0, 0, 0, 0.1),
-    0 1px 3px rgba(0, 0, 0, 0.08);
-  transition: box-shadow 0.3s ease;
-
-  &:hover {
-    box-shadow:
-      0 6px 8px rgba(0, 0, 0, 0.15),
-      0 3px 6px rgba(0, 0, 0, 0.1);
-  }
-`;
-
-const ButtonContent = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 6px;
-`;
+import { ButtonContent, SpoilerButton } from "../../ui/SpoilerButton";
 
 const ThinkingTextContainer = styled.span`
   display: inline-block;
